api/requirements.txt
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-text-splitters>=0.1.0
langchain_experimental>=0.1.0
langchain_huggingface>=0.1.0
fastapi>=0.104.0
loguru
azure-ai-formrecognizer>=3.3.3
uvicorn
chromadb

# For documentation
# sphinx==7.3.7
# sphinx_rtd_theme
# sphinx-autoapi